+++ dotnet/Services/VtexAPIService.cs
@@ -441,7 +441,7 @@ namespace Taxjar.Services
                         {
                             for (int lastTaxIndex = vtexTaxResponse.ItemTaxResponse[lastItemIndex].Taxes.Length - 1; lastTaxIndex >= 0; lastTaxIndex--)
                             {
-                                if (vtexTaxResponse.ItemTaxResponse[lastItemIndex].Taxes[lastTaxIndex].Value + adjustmentAmount >= 0)
+                                if (vtexTaxResponse.ItemTaxResponse[lastItemIndex].Taxes[lastTaxIndex].Value > 0 && vtexTaxResponse.ItemTaxResponse[lastItemIndex].Taxes[lastTaxIndex].Value + adjustmentAmount >= 0)
                                 {
                                     vtexTaxResponse.ItemTaxResponse[lastItemIndex].Taxes[lastTaxIndex].Value += adjustmentAmount;
                                     adjustmentAmount = 0;