+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Do not apply adjustment to zero value fields
+
 ## [0.1.10] - 2021-10-27
 
 ### Fixed