+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Front end validation for customer exemptions
+- Better error message when an invalid customer email is used
+
 ## [0.2.0] - 2021-11-10
 
 ### Added